перенес описание классов из main в src.classes_description
